requirements.txt
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
pluggy==1.5.0
pycparser==2.22
pypdfium2==4.30.0
pytest==8.3.3
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
six==1.16.0
tzdata==2024.2
